Fix boundsifpresent_fn when using AR1 structure in FieldConfig
New bounds were not matching the constraints of the 2 x 2 matrix,
for one category, when trying to set the bounds using boundsifpresent_fn,
which was passing a vector of length two for the upper and lower bounds.
The fix was tested for the single species example and appears to work.
I noticed the error when trying to implement the Triennial survey for the
West Coast.